Changing how related object name fields are discovered
In FieldSetService, when retrieving fields for a field set (or other
purposes), when preparing the data structure that contains details about
a particular field, if a lookup field is being represented the
referenced object name and the name of the field on the referenced
object that represents the name of the referenced object is determined
and included in the returned data structure representing the lookup
field.

The way determination of the name field of a referenced object was being
done was to issue a SOQL query against the FieldDefinition table.
However, when many fieldsets are being processed, this was causing this
SOQL query to be run for every lookup field that was included.  In some
cases, this was causing an exception in customer orgs because of "too
many SOQL queries".

To remediate this issue, this change replaces the logic of determining
the name field of an object type referenced by a lookup field with code
that will use the schema describe results within Apex to iterate over
fields of the referenced object and find the one that has the
'isNameField' property set to true.

This change is able to avoid governor limit exceptions for too many SOQL
queries, and does not seem to run into any additional limits even in
situations where many field sets with many lookups are defined in the
org (150 field sets with 4 lookup fields each, e.g.).